Add ActionHub to main content table
Thanks to the work Dan is mainly doing we now have a good amount of
actions. And more are coming as you can see from the open PR we have in
tinkerbell/hub repo.

I think it is time to officially link this initiative to the
documentation.

I think more about how to actually contribute can be covered as part of
the Contributor subsection of the documentation.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>