One sentence per line.
I know this is touching a lot of lines but I've got a bunch of content things I
want to fix and this just helps me go through all the prose. So in lieu of not
having any style guide and the mixed format a bunch of files had, I'd like to
"reset" everything to a single format for now. I'm not worried about any future
changes.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>